refactor: 출력 관련 로직을 Output클래스에 분리
